Start work on the project
I started work on the project and created table of content.
Until now I created 5 sections in the table of content.
I wrote the content of the sections up to section 3.1 inclusive.

In the next time I need to work and finish the sections 3.2-5.
In addition I need to try to start work on the models of the
project(Start from the Unet classic model with resize the img).