Works on Preprocessing part
I finished work on the third section and on the fourth section.
In addition, I worked on the fifth section and finished 5.1-5.4 sections inclusive .
In the next time I need to work about 5.5 section and finish him.
If I will have more time, so to start to work on 5.6 section.