Finished sections 7, 9-13
I finished work on sections 7, 9-13. In addition I learned how to work
with kaggle. I tryed to run basic model on him and I encouterd with problem, but after 3 days I solved this. I need to finish working on sections 6.2.7 and 6.3.
In addition I need to start work on section 8 and section 14, and start
to run the models on kaggle.